Variables

Variables (like e.g. CFVariable) are the quantities contained within a NetCDF dataset. See the Datasets page on how to obtain them from a dataset.

Different type of arrays are involved when working with NCDatasets. For instance assume that test.nc is a file with a Float32 variable called var. Assume that we open this data set in append mode ("a"):

using NCDatasets
ds = NCDataset("test.nc","a")
v_cf = ds["var"]

The variable v_cf has the type CFVariable. No data is actually loaded from disk, but you can query its size, number of dimensions, number elements, etc., using the functions size, ndims, length as if v_cf was an ordinary Julia array.

To load the variable v_cf in memory as numeric data you can convert it into an array (preserving its dimensionality structure) with

Array(v_cf)

The syntax v_cf[:] is equivalent with the above, it doesn't make a Vector (like it does on normal Julia arrays).

You can only load sub-parts of it in memory via indexing each dimension:

v_cf[1:5, 10:20]

(here you must know the number of dimensions of the variable, as you must access all of them).

The following functions are convenient for working with variables:

NCDatasets.dimnames — Function

dimnames(v::Variable)
dimnames(v::CFVariable)

Return a tuple of strings with the dimension names of the variable v.

NCDatasets.dimsize — Function

dimsize(v::CFVariable)

Get the size of a CFVariable as a named tuple of dimension → length.

NCDatasets.name — Function

name(v::Variable)

Return the name of the NetCDF variable v.

NCDatasets.nomissing — Function

a = nomissing(da)

Retun the values of the array da of type Array{Union{T,Missing},N} (potentially containing missing values) as a regular Julia array a of the same element type and checks that no missing values are present.

a = nomissing(da,value)

Retun the values of the array da of type Array{Union{T,Missing},N} as a regular Julia array a by replacing all missing value by value.

NCDatasets.loadragged — Function

 data = loadragged(ncvar,index::Colon)

Load data from ncvar in the contiguous ragged array representation as a vector of vectors. It is typically used to load a list of profiles or time series of different length each.

The indexed ragged array representation is currently not supported.

NCDatasets.load! — Function

load!(ncvar::Variable, data, indices)

Loads a NetCDF variables ncvar and puts the result in data along the specified indices.

data = zeros(5,6); # must have the right shape and type
load!(ds["temp"].var,data,:,:) # loads all data

data = zeros(5); # must have the right shape and type
load!(ds["temp"].var,data,:,1) # loads the 1st column

Dimensions

Dimensions are the dependent variables of a dataset (the ones with respect a Variable is defined). They are obtained with using dimnames on a Variable. We have the following functions relevant to them:

Base.setindex! — Method

Base.setindex!(d::Dimensions,len,name::AbstractString)

Defines the dimension called name to the length len. Generally dimension are defined by indexing, for example:

ds = NCDataset("file.nc","c")
ds.dim["longitude"] = 100

If len is the special value Inf, then the dimension is considered as unlimited, i.e. it will grow as data is added to the NetCDF file.

NCDatasets.unlimited — Method

unlimited(d::Dimensions)

Return the names of all unlimited dimensions.

Internals of a variable

NCDatasets.chunking — Function

storage,chunksizes = chunking(v::Variable)

Return the storage type (:contiguous or :chunked) and the chunk sizes of the varable v.

NCDatasets.deflate — Function

isshuffled,isdeflated,deflate_level = deflate(v::Variable)

Return compression information of the variable v. If shuffle is true, then shuffling (byte interlacing) is activaded. If deflate is true, then the data chunks (see chunking) are compressed using the compression level deflate_level (0 means no compression and 9 means maximum compression).

NCDatasets.checksum — Function

checksummethod = checksum(v::Variable)

Return the checksum method of the variable v which can be either be :fletcher32 or :nochecksum.

Creating a variable/dimension

defDim
defVar

Coordinate variables

NCDatasets.coord — Function

cv = coord(v::Union{CFVariable,Variable},standard_name)

Find the coordinate of the variable v by the standard name standard_name or some standardized heuristics based on units. If the heuristics fail to detect the coordinate, consider to modify the netCDF file to add the standard_name attribute. All dimensions of the coordinate must also be dimensions of the variable v.

Example

using NCDatasets
ds = NCDataset("file.nc")
ncv = ds["SST"]
lon = coord(ncv,"longitude")[:]
lat = coord(ncv,"latitude")[:]
v = ncv[:]
close(ds)