NCDatasets.jl

NCDatasets.jl

Documentation for NCDatasets.jl

Datasets

NCDatasets.Dataset — Type.

Dataset(filename::AbstractString,mode::AbstractString = "r";
                 format::Symbol = :netcdf4)

Create (mode = "c") or open in read-only (mode = "r") a NetCDF file (or an OPeNDAP URL). Supported formats:

:netcdf4 (default): HDF5-based NetCDF format

:netcdf4_classic: Only netCDF 3 compatible API features will be used

:netcdf3_classic: classic NetCDF format supporting only files smaller than 2GB.

:netcdf3_64bit_offset: improved NetCDF format supporting files larger than 2GB.

Files can also be open and automatically closed with a do block.

Dataset("file.nc") do ds
    data = ds["temperature"][:,:]
end 

source

Base.keys — Function.

keys(a::Attributes)

Return a list of the names of all attributes.

source

keys(ds::Dataset)

Return a list of all variables names in Dataset ds.

source

Base.haskey — Function.

haskey(ds::Dataset,varname)

Return true of the Dataset ds has a variable with the name varname. For example:

ds = Dataset("/tmp/test.nc","r")
if haskey(ds,"temperature")
    println("The file has a variable 'temperature'")
end

This example checks if the file /tmp/test.nc has a variable with the name temperature.

source

NCDatasets.variable — Function.

variable(ds::Dataset,varname::String)

Return the NetCDF variable varname in the dataset ds as a NCDataset.Variable. No scaling is applied when this variable is indexes.

source

NCDatasets.sync — Function.

sync(ds::Dataset)

Write all changes in Dataset ds to the disk.

source

Base.close — Function.

close(ds::Dataset)

Close the Dataset ds. All pending changes will be written to the disk.

source

Variables

NCDatasets.defVar — Function.

defVar(ds::Dataset,name,vtype,dimnames; kwargs...)

Define a variable with the name name in the dataset ds. vtype can be Julia types in the table below (with the corresponding NetCDF type). The parameter dimnames is a tuple with the names of the dimension. For scalar this parameter is the empty tuple (). The variable is returned (of the type CFVariable).

Keyword arguments

fillvalue: A value filled in the NetCDF file to indicate missing data. It will be stored in the _FillValue attribute.

chunksizes: Vector integers setting the chunk size.

The total size of a chunk must be less than 4 GiB.

deflatelevel: Compression level: 0 (default) means no compression

and 9 means maximum compression.

shuffle: If true, the shuffle filter is activated which can improve the

compression ratio.

checksum: The checksum method can be :fletcher32

or :nochecksum (checksumming is disabled, which is the default)

chunksizes, deflatelevel, shuffle and checksum can only be set on NetCDF 4 files.

NetCDF data types

NetCDF Type	Julia Type

NC_BYTE	Int8

NC_UBYTE	UInt8

NC_SHORT	Int16

NC_INT	Int32

NC_INT64	Int64

NC_FLOAT	Float32

NC_DOUBLE	Float64

NC_CHAR	Char

NC_STRING	String

source

NCDatasets.dimnames — Function.

dimnames(v::Variable)

Return a tuple of the dimension names of the variable v.

source

NCDatasets.name — Function.

name(v::Variable)

Return the name of the NetCDF variable v.

source

NCDatasets.chunking — Function.

storage,chunksizes = chunking(v::Variable)

Return the storage type (:contiguous or :chunked) and the chunk sizes of the varable v.

source

NCDatasets.deflate — Function.

shuffle,deflate,deflate_level = deflate(v::Variable)

Return compression information of the variable v. If shuffle is true, then shuffling (byte interlacing) is activaded. If deflate is true, then the data chunks (see chunking) are compressed using the compression level deflate_level (0 means no compression and 9 means maximum compression).

source

NCDatasets.checksum — Function.

checksummethod = checksum(v::Variable)

Return the checksum method of the variable v which can be either be :fletcher32 or :nochecksum.

source

Base.start — Method.

start(ds::Dataset)

for (varname,var) in ds
    @show (varname,size(var))
end

source

Attributes

The NetCDF dataset (as return by Dataset) and the NetCDF variables (as returned by getindex, variable or defVar) have the field attrib which has the type NCDatasets.Attributes and behaves like a julia dictionary.

Base.getindex — Method.

getindex(a::Attributes,name::AbstractString)

Return the value of the attribute called name from the attribute list a. Generally the attributes are loaded by indexing, for example:

ds = Dataset("file.nc")
title = ds.attrib["title"]

source

Base.setindex! — Method.

Base.setindex!(a::Attributes,data,name::AbstractString)

Set the attribute called name to the value data in the attribute list a. Generally the attributes are defined by indexing, for example:

ds = Dataset("file.nc","c")
ds.attrib["title"] = "my title"

source

Base.keys — Method.

keys(a::Attributes)

Return a list of the names of all attributes.

source

Dimensions

NCDatasets.defDim — Function.

defDim(ds::Dataset,name,len)

Define a dimension in the data-set ds with the given name and length len. If len is the special value Inf, then the dimension is considered as unlimited, i.e. it will grow as data is added to the NetCDF file.

For example:

ds = Dataset("/tmp/test.nc","c")
defDim(ds,"lon",100)

This defines the dimension lon with the size 100.

source

Base.setindex! — Method.

Base.setindex!(d::Dimensions,len,name::AbstractString)

Defines the dimension called name to the length len. Generally dimension are defined by indexing, for example:

ds = Dataset("file.nc","c")
ds.dim["longitude"] = 100

If len is the special value Inf, then the dimension is considered as unlimited, i.e. it will grow as data is added to the NetCDF file.

source