Experimental features

Multi-file support

Multiple files can also be aggregated over a given dimensions (or the record dimension). In this example, 3 sea surface temperature fields from the 1992-01-01 to 1992-01-03 are aggregated using the OPeNDAP service from PODAAC.

using NCDatasets, Printf, Dates

function url(dt)
  doy = @sprintf("%03d",Dates.dayofyear(dt))
  y = @sprintf("%04d",Dates.year(dt))
  yyyymmdd = Dates.format(dt,"yyyymmdd")
  return "https://podaac-opendap.jpl.nasa.gov:443/opendap/allData/ghrsst/data/GDS2/L4/GLOB/CMC/CMC0.2deg/v2/$y/$doy/$(yyyymmdd)120000-CMC-L4_GHRSST-SSTfnd-CMC0.2deg-GLOB-v02.0-fv02.0.nc"
end

ds = NCDataset(url.(DateTime(1992,1,1):Dates.Day(1):DateTime(1992,1,3)),aggdim = "time");
SST2 = ds["analysed_sst"][:,:,:];
close(ds)

If there is a network or server issue, you will see an error message like NetCDF: I/O failure.

Data selection based on values

NCDatasets.@select — Macro

vsubset = NCDatasets.@select(v,expression)

Return a subset of the variable v following the condition expression as a view. The condition has the following form:

condition1 && condition2 && condition3 ... conditionN

Every condition should involve a single 1D NetCDF variable (typically a coordinate variable, referred as coord below) and should have a shared dimension with the variable v. All local variables (including local functions) need to have a $ prefix (see examples below) as the condition is evaluated in the scope of the module NCDatasets[1]. This macro is highly experimental and subjected to change.

Every condition can either perform:

a nearest match: coord ≈ target_coord. Only the data corresponding to the index closest to target_coord is loaded.

a nearest match with tolerance: coord ≈ target_coord ± tolerance. As before, but if the difference between the closest value in coord and target_coord is larger (in absolute value) than tolerance, an empty array is returned.

a condition operating on scalar values. For example, a condition equal to 10 <= lon <= 20 loads all data with the longitude between 10 and 20 or abs(lat) > 60 loads all variables with a latitude north of 60° N and south of 60° S (assuming that the NetCDF has the 1D variables lon and lat for longitude and latitude).

Only the data which satisfies all conditions is loaded. All conditions must be chained with an && (logical and). They should not contain additional parenthesis or other logical operators such as || (logical or).

To convert the view into a regular array one can use collect or Array for arrays. As in julia, views of scalars are wrapped into a zero dimensional arrays which can be dereferenced by using []. Modifying a view will modify the underlying NetCDF file (if the file is opened as writable).

As for any view, one can use parentindices(vsubset) to get the indices matching a select query.

Examples

Create a sample file with random data:

using NCDatasets, Dates
fname = "sample_file.nc"
lon = -180:180
lat = -90:90
time = DateTime(2000,1,1):Day(1):DateTime(2000,1,3)
SST = randn(length(lon),length(lat),length(time))

ds = NCDataset(fname,"c")
defVar(ds,"lon",lon,("lon",));
defVar(ds,"lat",lat,("lat",));
defVar(ds,"time",time,("time",));
defVar(ds,"SST",SST,("lon","lat","time"));

# load by bounding box
v = NCDatasets.@select(ds["SST"],30 <= lon <= 60 && 40 <= lat <= 90)

# substitute a local variable in condition using $
lonr = (30,60) # longitude range
latr = (40,90) # latitude range

v = NCDatasets.@select(v,$lonr[1] <= lon <= $lonr[2] && $latr[1] <= lat <= $latr[2])

# get the indices matching a select query
(lon_indices,lat_indices,time_indices) = parentindices(v)

# find the nearest time instance
v = NCDatasets.@select(ds["SST"],time ≈ DateTime(2000,1,4))

# find the nearest time instance but not earlier or later than 2 hours
# an empty array is returned if no time instance is present

v = NCDatasets.@select(ds["SST"],time ≈ DateTime(2000,1,3,1) ± Hour(2))

# Note DateTime and Hour do not need to a $ prefix because NCDataset imports
# the modules Dates and CFTime.

close(ds)

Any 1D variable with the same dimension name can be used in @select. For example, if we have a time series of temperature and salinity, the temperature values can also be selected based on salinity:

# create a sample time series
using NCDatasets, Dates
fname = "sample_series.nc"
time = DateTime(2000,1,1):Day(1):DateTime(2009,12,31)
salinity = randn(length(time)) .+ 35
temperature = randn(length(time))

NCDataset(fname,"c") do ds
    defVar(ds,"time",time,("time",));
    defVar(ds,"salinity",salinity,("time",));
    defVar(ds,"temperature",temperature,("time",));
end

ds = NCDataset(fname)

# load all temperature data from January where the salinity is larger than 35.
v = NCDatasets.@select(ds["temperature"],Dates.month(time) == 1 && salinity >= 35)

# this is equivalent to
v2 = ds["temperature"][findall(Dates.month.(time) .== 1 .&& salinity .>= 35)]

@test v == v2
close(ds)

Note

For optimal performance, one should try to load contiguous data ranges, in particular when the data is loaded over HTTP/OPeNDAP.

Experimental functions

NCDatasets.ancillaryvariables — Function

ncvar = NCDatasets.ancillaryvariables(ncv::NCDatasets.CFVariable,modifier)

Return the first ancillary variables from the NetCDF variable ncv with the standard name modifier modifier. It can be used for example to access related variable like status flags.

Base.filter — Function

data = NCDatasets.filter(ncv, indices...; accepted_status_flags = nothing)

Load and filter observations by replacing all variables without an acepted status flag to missing. It is used the attribute ancillary_variables to identify the status flag.

# da["data"] is 2D matrix
good_data = NCDatasets.filter(ds["data"],:,:, accepted_status_flags = ["good_data","probably_good_data"])

1Please file an issue if you know a better solution.